requirement.txt
python-telegram-bot==13.15
APScheduler==3.10.4
Pillow==10.2.0
requests
pytz
flask
schedule